requirements.txt
streamlit==1.31.0
snowflake-connector-python
pandas
plotly==5.18.0
snowflake-ml-python
numpy
requests